improvements

+++ club/src/main/java/com/qr/club/service/QrService.java
@@ -2,6 +2,7 @@ package com.qr.club.service;
 
 import com.qr.club.entity.Member;
 import com.qr.club.entity.dto.MemberDto;
+import com.qr.club.entity.response.MemberResponse;
 import com.qr.club.exception.MemberRepositoryException;
 import com.qr.club.mapper.MemberMapper;
 import com.qr.club.repository.QrRepository;
@@ -20,14 +21,14 @@ public class QrService {
     private final QrRepository repository;
     private final MemberMapper mapper;
 
-    public Member create(MemberDto dto) {
-        return repository.save(
+    public MemberResponse create(MemberDto dto) {
+        Member member = repository.save(
                 mapper.toMember(dto)
         );
-
+        return mapper.toMemberResponse(member);
     }
 
-    public MemberDto update(MemberDto dto, UUID uuid) {
+    public MemberResponse update(MemberDto dto, UUID uuid) {
 
         Member member = repository.findOptionalByUuid(uuid)
                 .orElseThrow(() -> new MemberRepositoryException("Member not found"));
@@ -39,21 +40,26 @@ public class QrService {
         Optional.ofNullable(dto.middleName())
                 .ifPresent(member::setMiddleName);
 
-        repository.save(member);
-
-        return mapper.toMemberDto(member);
+        return mapper.toMemberResponse(
+                repository.save(member)
+        );
     }
 
     public void deleteByUuid(UUID uuid) {
-        repository.deleteByUuid(uuid);
+        if (repository.existsByUuid(uuid)) {
+            repository.deleteByUuid(uuid);
+        } else {
+            throw new MemberRepositoryException("Member not found");
+        }
     }
 
-    public MemberDto enterClub(UUID uuid) {
+    public MemberResponse enterClub(UUID uuid) {
        Member member = repository.findOptionalByUuid(uuid)
                .orElseThrow(() -> new MemberRepositoryException("Member not found"));
        member.setUuid(UUID.randomUUID());
-       repository.save(member);
 
-       return mapper.toMemberDto(member);
+       return mapper.toMemberResponse(
+               repository.save(member)
+       );
     }
 }
